Nest Custom views (Todo Within Task) - to further seperation of concerns
  -Created TaskView.kt - Kotlin
  -Implement Callbacks
      - (invokes)TodoView.kt
      - (implements)TaskView.kt

Conceptually a very abstract commit, implements annoymouse functions similar to that of javascript, to check if todos are completed so that we are able to strike through the titleView Text.

Invoke annoymouse functions when any todo checkbox has an interaction, Inheritence will scope the todos for us.

TaskView.kt
  - Now implements TaskAdapter logic, to further seperation of concerns
  - Invokes Callback
  - Interates through list of todos (from the scope of TodoView) to check if all todos are checked
  - Strikes TitleView if todos are all completed

TodoView.kt
  - Callback Implementation
    - Initially passed to initView()
    - Invoked within setupCheckStateListener()
    - Scope, allows todo within this context to use functions from TaskView.kt